Recursion
This is a recursion that allows a function to use itself instead of using a loops which cuts down on processing time for large amounts of data. this program uses recursion to find the largest number by looping through itself by comparing each number until it gets to the last one.